+++ packages/preview-server/scripts/utils/default-seed/communications/webhooks-failed.tsx
@@ -55,10 +55,10 @@ export default function WebhooksFailed({ date, error }: WebhooksFailedProps) {
               If you have any questions or need assistance, please reach out to
               us at{' '}
               <a
-                href="mailto:support@useprint.app"
+                href="mailto:support@useprint.dev"
                 className="text-cyan-300 underline"
               >
-                support@useprint.app
+                support@useprint.dev
               </a>
               .
             </Text>