+++ packages/preview-server/src/actions/document-validation/check-images.spec.tsx
@@ -6,7 +6,7 @@ test('checkImages()', async () => {
   <img src="https://useprint.com/static/brand/useprint-icon-white.png" />,
   <img src="/static/codepen-challengers.png" alt="codepen challenges" />,
 </div>`;
-  const stream = await checkImages(html, 'https://demo.useprint.app');
+  const stream = await checkImages(html, 'https://demo.useprint.dev');
   const reader = stream.getReader();
   while (true) {
     const { done, value } = await reader.read();