+++ packages/render/src/browser/render.tsx
@@ -1,9 +1,5 @@
-import { convert } from 'html-to-text';
 import { Suspense } from 'react';
 import type { ReactDOMServerReadableStream } from 'react-dom/server';
-import { pretty } from '../node';
-import type { Options } from '../shared/options';
-import { plainTextSelectors } from '../shared/plain-text-selectors';
 
 const decoder = new TextDecoder('utf-8');
 
@@ -38,7 +34,7 @@ const readStream = async (stream: ReactDOMServerReadableStream) => {
   return decoder.decode(mergedChunks);
 };
 
-export const render = async (node: React.ReactNode, options?: Options) => {
+export const render = async (node: React.ReactNode) => {
   const suspendedElement = <Suspense>{node}</Suspense>;
   const reactDOMServer = await import('react-dom/server.browser').then(
     // This is beacuse react-dom/server is CJS
@@ -57,21 +53,10 @@ export const render = async (node: React.ReactNode, options?: Options) => {
       .catch(reject);
   });
 
-  if (options?.plainText) {
-    return convert(html, {
-      selectors: plainTextSelectors,
-      ...options.htmlToTextOptions,
-    });
-  }
 
   const doctype =
     '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">';
 
   const document = `${doctype}${html.replace(/<!DOCTYPE.*?>/, '')}`;
-
-  if (options?.pretty) {
-    return pretty(document);
-  }
-
   return document;
 };