+++ packages/render/src/node/index.ts
@@ -1,14 +1,12 @@
-import type { Options } from '../shared/options';
+
 import { render } from './render';
 
 /**
  * @deprecated use {@link render}
  */
-export const renderAsync = (element: React.ReactElement, options?: Options) => {
-  return render(element, options);
+export const renderAsync = (element: React.ReactElement) => {
+  return render(element);
 };
 
-export * from '../shared/options';
 export * from '../shared/plain-text-selectors';
-export * from '../shared/utils/pretty';
 export * from './render';