+++ packages/render/src/node/render.tsx
@@ -1,11 +1,7 @@
-import { convert } from 'html-to-text';
 import { Suspense } from 'react';
-import type { Options } from '../shared/options';
-import { plainTextSelectors } from '../shared/plain-text-selectors';
-import { pretty } from '../shared/utils/pretty';
 import { readStream } from './read-stream';
 
-export const render = async (node: React.ReactNode, options?: Options) => {
+export const render = async (node: React.ReactNode) => {
   const suspendedElement = <Suspense>{node}</Suspense>;
   const reactDOMServer = await import('react-dom/server').then(
     // This is beacuse react-dom/server is CJS
@@ -31,21 +27,11 @@ export const render = async (node: React.ReactNode, options?: Options) => {
     });
   }
 
-  if (options?.plainText) {
-    return convert(html, {
-      selectors: plainTextSelectors,
-      ...options.htmlToTextOptions,
-    });
-  }
 
   const doctype =
     '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">';
 
   const document = `${doctype}${html.replace(/<!DOCTYPE.*?>/, '')}`;
-
-  if (options?.pretty) {
-    return pretty(document);
-  }
 
   return document;
 };